:recycle: refactor(SellerControllerTest): tests refactored so that it delets Sellers saved at the end of tests
