+++ src/test/java/com/abneco/delivery/user/mock/MockSellerRepository.java
@@ -18,7 +18,6 @@ public class MockSellerRepository implements SellerRepository {
         return Optional.empty();
     }
 
-
     public Optional<Seller> findByEmail() {
         return Optional.of(new Seller());
     }
